# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Baselix API
  version: 1.0.0
  description: |
    Baselix is a dynamic data-storage API built on an Entity–Attribute–Value (EAV) model.
    You define tables, manage their schemas, and perform full CRUD operations on records —
    all through a simple REST API.

    ## Authentication

    Every API request must include a project API key as a Bearer token:

    ```
    Authorization: Bearer <api_key>
    ```

    API keys are generated per project and can be rotated from the dashboard.

    ## Field Types

    When defining a schema, each field is assigned a type string:

    | Type | Description |
    |------|-------------|
    | `string` | Text |
    | `int` | Integer |
    | `float` | Decimal number |
    | `bool` | Boolean |
    | `time` | ISO 8601 timestamp |
    | `json` | JSON object or array |
    | `uuid` | UUID string |

    Append `_u` to any type (e.g. `string_u`) to enforce a **uniqueness constraint** on that field.
    Only `string`, `int`, `float`, and `uuid` support the `_u` suffix.

    ## Filtering & Sorting

    ### Filter syntax
    ```
    ?filter=field:operator:value,field:operator:value
    ```

    Supported operators: `eq`, `ne`, `gt`, `gte`, `lt`, `lte`, `contains`

    You can filter on user-defined fields as well as the meta fields `id`, `created_at`, and `updated_at`.
    Use `null` as the value for null checks (e.g. `?filter=name:eq:null`).

    ### Sort syntax
    ```
    ?sort=field:direction,field:direction
...

...
servers:
...
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: integer
          example: 400
        message:
          type: string
          example: "table 'users' already exists"
    SchemaObject:
      type: object
      description: >
        A map of field names to their types.
        Append `_u` to a type to make the field unique (e.g. `"email": "string_u"`).
        Reserved names `id`, `created_at`, and `updated_at` are not allowed.
      additionalProperties:
        type: string
      example:
        name: string_u
        age: int
        score: float
        is_active: bool
        joined_at: time
        metadata: json
        ref_id: uuid
    TableFields:
      type: object
      description: Ordered map of field names to their type strings.
      additionalProperties:
        type: string
      example:
        name: string_u
        age: int
    TableSchema:
      type: object
      properties:
        name:
          type: string
          example: users
        fields:
          $ref: "#/components/schemas/TableFields"
...
paths:
...
  /tables:
    get:
      summary: List all tables
      description: Returns all tables and their schemas for the authenticated project.
      operationId: getTables
      tags: [Tables]
      responses:
        "200":
          description: List of tables
          content:
            application/json:
              schema:
                type: object
                properties:
                  tables:
                    type: array
                    items:
                      $ref: "#/components/schemas/TableSchema"
              example:
                tables:
                  - name: users
                    fields:
                      name: string_u
                      age: int
                  - name: products
                    fields:
                      title: string
                      price: float
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: 401
                message: "invalid API key"
  /tables/{name}/schema:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
        description: Table name
        example: users

    get:
      summary: Get table schema
      description: Returns the schema of a single table.
      operationId: getTable
      tags: [Tables]
      responses:
        "200":
          description: Table schema
          content:
            application/json:
              schema:
                type: object
                properties:
                  tables:
                    $ref: "#/components/schemas/TableSchema"
              example:
                tables:
                  name: users
                  fields:
                    name: string_u
                    age: int
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Table not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: 404
                message: "table not found"

    post:
      summary: Create a table
      description: >
        Creates a new table with the given schema.
        Returns an error if a table with that name already exists for this project.
      operationId: createTable
      tags: [Tables]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [schema]
              properties:
                schema:
                  $ref: "#/components/schemas/SchemaObject"
            example:
              schema:
                name: string_u
                age: int
                score: float
      responses:
        "200":
          description: Table created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  table_id:
                    type: string
                    format: uuid
                  table_name:
                    type: string
              example:
                message: "Table 'users' created successfully for project 'my-project'"
                table_id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                table_name: users
        "400":
          description: Table already exists, reserved field name, or invalid schema
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                already_exists:
                  summary: Table already exists
                  value:
                    error: 400
                    message: "table 'users' already exists"
                reserved_field:
                  summary: Reserved field name
                  value:
                    error: 400
                    message: "cannot use reserved field: id, we manage these fields automatically for you: id, created_at, updated_at"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Update a table schema
      description: |
        Updates the schema for an existing table. **This is a destructive operation:**

        - Fields removed from the new schema will be **deleted along with all their stored values**.
        - Fields whose type changes will have all their **existing values deleted**.
        - New fields are added to the schema (existing records will have `null` for those fields).

        You must include `"allow_destructive": true` in the request body to confirm the operation.
      operationId: updateTable
      tags: [Tables]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [schema, allow_destructive]
              properties:
                schema:
                  $ref: "#/components/schemas/SchemaObject"
                allow_destructive:
                  type: boolean
                  description: Must be `true` to permit destructive changes.
            example:
              schema:
                name: string_u
                age: int
                email: string_u
              allow_destructive: true
      responses:
        "200":
          description: Table updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  table_id:
                    type: string
                    format: uuid
                  table_name:
                    type: string
              example:
                message: "Table 'users' updated successfully for project 'my-project'"
                table_id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                table_name: users
        "400":
          description: "`allow_destructive` not set to `true`, reserved field name, or invalid schema"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: 400
                message: "Warning, this route will perform destructive changes (refer to the API documentation for more details). Destructive changes are not allowed without 'allow_destructive' flag set to true"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete a table
      description: >
        Permanently deletes a table and **all** of its associated data
        (fields, records, and values). This action is irreversible.
      operationId: deleteTable
      tags: [Tables]
      responses:
        "200":
          description: Table deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "Table 'users' deleted successfully for project 'my-project'"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Table not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: 404
                message: "table not found"

  # ─────────────────────────────────────────────
  # Records
  # ─────────────────────────────────────────────
...